xianyubot/requirements.txt
httpx==0.28.1
langchain_openai==1.1.11
lightrag_hku==1.4.10
loguru==0.7.3
msgpack_python==0.5.6
numpy==2.4.3
openai==2.26.0
python-dotenv==1.2.2
Requests==2.32.5
sentence_transformers==5.3.0
websockets==16.0